+++ Documentation~/whats-new.md
@@ -10,9 +10,15 @@ For a complete list of changes made, refer to the **Changelog** page.
 
 The main updates in this release include:
 
-## [1.0.1] - 2024-02-14
+## [1.0.2] - 2024-02-26
 
 ### Fixed
 
-- Fixed ColorSwatch shader to support Shader Model older than 5.0
+- Fixed the clipping flag on the GridView's ScrollView content container to prevent overscroll on touch devices.
+- Fixed unit tests
+- Disabled the dragging feature in the GridView when used on touch devices.
+
+### Added
+
+- Improved Localization Unity Package support with `LangContext` propagation.